MB-28737: UBSan: Fix integer overflows
UBSan flags a number of cases where we overflow an integer type - for
example:

    signed integer overflow: 1073741824 * 1024 cannot be represented in type 'int'
    #0 0x1029596c0 in parse_config(char const*, config_item*, __sFILE*) config_parser.cc:131

Reviewed-on: http://review.couchbase.org/91017
Tested-by: Build Bot <build@couchbase.com>

+++ utilities/config_parser.cc
@@ -122,8 +122,8 @@ int parse_config(const char *str, struct config_item *items, FILE *error) {
             case DT_SSIZE:
                {
                   const char *sfx = "kmgt";
-                  int multiplier = 1;
-                  int m = 1;
+                  size_t multiplier = 1;
+                  size_t m = 1;
                   const char *p;
 
                   for (p = sfx; *p != '\0'; ++p) {
